added fault and recovery times as flags to all modes

+++ FaultInjector.py
@@ -351,7 +351,7 @@ class Ceph(Fault):
         """
         print 'Beginning Ceph Stateful Mode...\n'
 
-        thread_count = self.deployment.min_replication_size + math.ceil(self.deployment.num_mons / 2)
+        thread_count = int(self.deployment.min_replication_size + math.ceil(self.deployment.num_mons / 2))
 
         # create threads and append them to the global thread list and fault thread list
         fault_threads = []
